refactor: 폴더 구조 맞게 수정
