+++ src/pages/Shop/components/UnOrderableComponent.tsx
@@ -1,15 +1,15 @@
 import { useRef } from 'react';
 import { useParams } from 'react-router-dom';
-import Header from './components/Header';
-import ImageCarousel from './components/ImageCarousel';
-import ShopMenuGroups from './components/ShopMenuGroups';
-import ShopMenus from './components/ShopMenus';
-import ShopSummary from './components/ShopSummary';
-import { useGetUnOrderableShopInfo } from './hooks/useGetShopInfo';
-import { useGetUnOrderableShopReviews } from './hooks/useGetShopInfo';
-import { useGetUnOrderableShopMenuGroups } from './hooks/useGetShopInfo';
-import { useGetUnOrderableShopMenus } from './hooks/useGetShopInfo';
-import { useInteraction } from './hooks/useInteraction';
+import { useGetUnOrderableShopInfo } from '../hooks/useGetShopInfo';
+import { useGetUnOrderableShopReviews } from '../hooks/useGetShopInfo';
+import { useGetUnOrderableShopMenuGroups } from '../hooks/useGetShopInfo';
+import { useGetUnOrderableShopMenus } from '../hooks/useGetShopInfo';
+import { useInteraction } from '../hooks/useInteraction';
+import Header from './Header';
+import ImageCarousel from './ImageCarousel';
+import ShopMenuGroups from './ShopMenuGroups';
+import ShopMenus from './ShopMenus';
+import ShopSummary from './ShopSummary';
 
 interface UnOrderableComponentProps {
   totalQuantity: number;
